Re-organize logging code a bit - move custom filters and handlers into st2common.logging package.
