+++ st2common/st2common/log.py
@@ -15,17 +15,20 @@
 
 from __future__ import absolute_import
 
-import socket
-import datetime
 import logging
 import logging.config
 import logging.handlers
-import os
-import six
 import sys
 import traceback
 
+import six
 from oslo.config import cfg
+
+from st2common.logging.filters import ExclusionFilter
+
+# Those are here for backward compatibility reasons
+from st2common.logging.handlers import FormatNamedFileHandler
+from st2common.logging.handlers import ConfigurableSyslogHandler
 
 __all__ = [
     'getLogger',
@@ -33,9 +36,6 @@ __all__ = [
 
     'FormatNamedFileHandler',
     'ConfigurableSyslogHandler',
-
-    'ExclusionFilter',
-    'LogLevelFilter',
 
     'LoggingStream'
 ]
@@ -47,65 +47,6 @@ logging.addLevelName(logging.AUDIT, 'AUDIT')
 def getLogger(name):
     logger_name = 'st2.{}'.format(name)
     return logging.getLogger(logger_name)
-
-
-class FormatNamedFileHandler(logging.FileHandler):
-    def __init__(self, filename, mode='a', encoding=None, delay=False):
-        # Include timestamp in the name.
-        filename = filename.format(ts=str(datetime.datetime.utcnow()).replace(' ', '_'),
-                                   pid=os.getpid())
-        super(FormatNamedFileHandler, self).__init__(filename, mode, encoding, delay)
-
-
-class ConfigurableSyslogHandler(logging.handlers.SysLogHandler):
-    def __init__(self, address=None, facility=None, socktype=None):
-        if not address:
-            address = (cfg.CONF.syslog.host, cfg.CONF.syslog.port)
-        if not facility:
-            facility = cfg.CONF.syslog.facility
-        if not socktype:
-            protocol = cfg.CONF.syslog.protocol.lower()
-
-            if protocol == 'udp':
-                socktype = socket.SOCK_DGRAM
-            elif protocol == 'tcp':
-                socktype = socket.SOCK_STREAM
-            else:
-                raise ValueError('Unsupported protocol: %s' % (protocol))
-
-        if socktype:
-            super(ConfigurableSyslogHandler, self).__init__(address, facility, socktype)
-        else:
-            super(ConfigurableSyslogHandler, self).__init__(address, facility)
-
-
-class ExclusionFilter(object):
-
-    def __init__(self, exclusions):
-        self._exclusions = set(exclusions)
-
-    def filter(self, record):
-        if len(self._exclusions) < 1:
-            return True
-        module_decomposition = record.name.split('.')
-        exclude = len(module_decomposition) > 0 and module_decomposition[0] in self._exclusions
-        return not exclude
-
-
-class LogLevelFilter(logging.Filter):
-    """
-    Filter which excludes log messages which match the provided log levels.
-    """
-
-    def __init__(self, log_levels):
-        self._log_levels = log_levels
-
-    def filter(self, record):
-        level = record.levelno
-        if level in self._log_levels:
-            return False
-
-        return True
 
 
 class LoggingStream(object):
